feat(auth): 添加角色权限缓存功能
- 在 TokenService 中添加角色权限信息的缓存逻辑- 新增 RedisConfig 配置类，用于配置 RedisTemplate
- 新增 RedisService 服务类，提供 Redis 操作方法
- 更新 RolePermissionsDTO 中 roleId 类型从 Integer 改为 Long

+++ xh-auth/src/main/java/com/xh/auth/service/dto/RolePermissionsDTO.java
@@ -15,7 +15,7 @@ import java.util.List;
 @Data
 public class RolePermissionsDTO implements Serializable {
     @Schema(title = "角色ID")
-    protected Integer roleId;
+    protected Long roleId;
     @Schema(title = "创建时间")
     private LocalDateTime createTime;
     @Schema(title = "修改时间")
